renames; rename dht_op -> dht_op_light
